Start extraction of chart from chart completer
