+++ src/opennlp/ccg/parse/Chart.java
@@ -1,0 +1,5 @@
+package opennlp.ccg.parse;
+
+public interface Chart {
+
+}
